Add JedisPool to Java quickstart
For the Java Jedis quickstart, use `JedisPool` to provide the `Jedis`
instance so that the app is ready to scale up to multiple threads. The
general setup for `JedisPool` comes from a previously published
[gist][0].

This aligns the quickstart more closely with the ["Basic usage
example"][1] from the upstream Jedis docs.

Other changes:

Move a couple of punctuation characters to match the style used in the
Jedis source code.

[0]: https://gist.github.com/JonCole/925630df72be1351b21440625ff2671f#file-redis-samplecode-java-jedispool-java
[1]: https://github.com/redis/jedis/wiki/Getting-started/cd6eb0cebe4569291a1fc814192515ce6bd1c01c#basic-usage-example